+++ Code/Commando/dlgloadspgame.cpp
@@ -369,7 +369,7 @@ LoadSPGameMenuClass::On_ListCtrl_Column_Click
 		if (col_index == CurrSortCol) {
 			IsSortAscending = !IsSortAscending;
 		} else {
-			CurrSortCol = col_index;
+			CurrSortCol = uint16(col_index);
 			IsSortAscending = true;
 		}
